Rename the hashing module to authnext (authenticated non-existence) as NSEC doesn't do any hashing, only NSEC3 does.
