--- src/sign/hashing/mod.rs
@@ -1,3 +1,0 @@
-pub mod config;
-pub mod nsec;
-pub mod nsec3;
