+++ src/sign/signing/traits.rs
@@ -18,7 +18,7 @@ use crate::base::record::Record;
 use crate::base::Name;
 use crate::rdata::ZoneRecordData;
 use crate::sign::error::{SignError, SigningError};
-use crate::sign::hashing::nsec3::Nsec3HashProvider;
+use crate::sign::authnext::nsec3::Nsec3HashProvider;
 use crate::sign::keys::keymeta::DesignatedSigningKey;
 use crate::sign::records::{
     DefaultSorter, RecordsIter, Rrset, SortedRecords, Sorter,
